fix(portfolio): seed Workforce taxonomy root label
